% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_utils.R
\name{md5table}
\alias{md5table}
\title{md5table
For use in data analysis provenance, generates a data frame with file names and md5 values for use in checking for changes to source data files.  Is not able to validate changes to source data files.}
\usage{
md5table(file_list)
}
\arguments{
\item{file_list}{single file or vector of file names used as input}
}
\value{
data_frame with file name and md5 value
}
\description{
md5table
For use in data analysis provenance, generates a data frame with file names and md5 values for use in checking for changes to source data files.  Is not able to validate changes to source data files.
}
\examples{
\dontrun{
md5table("file_list")
}
}

